+++ __tests__/Car.test.js
@@ -1,4 +1,4 @@
-const Car = require('../src/Car');
+import Car from '../src/domain/Car.js';
 
 describe('Car.move', () => {
   test('move 메서드를 실행하면, position필드 값이 1 증가하여야 한다.', () => {
