+++ __tests__/RacingGame.test.js
@@ -1,4 +1,4 @@
-const RacingGame = require('../src/RacingGame');
+import RacingGame from '../src/domain/RacingGame.js';
 
 describe('RacingGame 메소드들이 예상 값을 반환하는 지 테스트', () => {
   const names = ['a', 'b'];
