+++ src/view/InputView.js
@@ -1,4 +1,4 @@
-const Console = require('../utils/Console');
+import Console from '../utils/Console.js';
 
 const REQUEST_MSG = {
   carNames: '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n',
@@ -15,4 +15,4 @@ const InputView = {
   },
 };
 
-module.exports = InputView;
+export default InputView;
