+++ src/view/OutputView.js
@@ -1,4 +1,4 @@
-const { Converter } = require('../utils');
+import { Converter } from '../utils/index.js';
 
 const OutputView = {
   printResultMessage() {
@@ -29,4 +29,4 @@ const RESULT_MESSAGE = {
   ending: '(이)가 최종 우승했습니다.',
 };
 
-module.exports = OutputView;
+export default OutputView;
